requirements.txt
tensorflow
pandas
opencv-python
pyglet
pillow
